refactor: mint version of gen_*

+++ src/code/math/cantor_seq2val.hpp
@@ -11,7 +11,7 @@ inline u64 cantor_seq2val(vec<u32> const &s, vec<u64> const &fact, u64 mod = (u6
   for (u64 i = 0; i < n; ++i) (ret += mul_mod_u(s[i], fact[n - i - 1], mod)) %= mod;
   return ret;
 }
-inline u64 cantor_seq2val(vec<u32> const &s, u64 mod = (u64)-1) { return cantor_seq2val(s, gen_fact(s.size(), mod), mod); }
+inline u64 cantor_seq2val(vec<u32> const &s, u64 mod = (u64)-1) { return cantor_seq2val(s, gen_fact((u32)s.size(), mod), mod); }
 
 }  // namespace tifa_libs::math
